+++ CHANGELOG.md
@@ -21,7 +21,13 @@
   - [#1627](https://github.com/wix-incubator/rich-content/pull/1627) added `ricos-editor-modal` class to the editor modal
 
   </details>
+
+### :rocket: New Feature
+- `ricos-content`
+  - [#1622](https://github.com/wix-incubator/rich-content/pull/1622) `isContentStateEmpty` util
+  - 
 <hr/>
+
 
 ## 8.0.0 TSLA5000 (Oct 12, 2020)
 ### :rocket: New Feature